// Uni_MS/angular-ui/src/app/models/teacher.ts
export interface Teacher {
  id?: number;
  uniqueCode?: string;
  teacherCode: string;
  firstName: string;
  lastName: string;
  gender?: string;
  dateOfBirth?: string;
  bloodGroup?: string;
  nationalId?: string;
  passport?: string;
  nationality?: string;
  religion?: string;
  maritalStatus?: string;
  photo?: string;
  email: string;
  phone?: string;
  emergencyContact?: string;
  presentAddress?: string;
  permanentAddress?: string;
  joiningDate?: string;
  employmentStatus?: string;
  employmentType?: string;
  designation?: string;
  departmentId?: number;
  facultyId?: number;
  officeRoom?: string;
  campus?: string;
  highestDegree?: string;
  university?: string;
  specialization?: string;
  experience?: string;
  certifications?: string;
  assignedCourses?: string;
  sections?: string;
  semester?: string;
  creditLoad?: string;
  googleScholar?: string;
  orcid?: string;
  salaryGrade?: string;
  basicSalary?: number;
  bankInformation?: string;
  taxId?: string;
  status?: string;
  userId?: number;
  createdAt?: string;
  updatedAt?: string;
}